Break apart the Etcd client
The Etcd Client is a little tricker because it requires
a config file. Later when we unify all the clients we will
use individual functions just like etcd.
